cleaned up testing module files, improved mock paths

+++ tests/modules/test_download/test_download.py
@@ -2,39 +2,45 @@ import unittest
 from unittest import mock
 from pathlib import Path
 from shutil import rmtree
-import os
-
-# import mocks
-# from tests.mocks.driver import initialize_test_driver
+from dotenv import load_dotenv
 
 # project modules
 from modules.download import download_pdf
-from locations import dirs, paths
-from logs.config.logging import logs_config
 from modules.initialize import initialize_driver
+from locations import dirs, paths, root_dir, test_dir
+from logs.config.logging import logs_config
 
-# MOCKS
+# LOGGING
+logs_config(paths["logs_config_test"])
+
+# ENV
+load_dotenv(root_dir / ".dev.env")
+
+# MOCK VARS
 mock_dirs = {
-    "pdfs": Path("../../output/pdfs/").resolve()
+    "pdfs": test_dir / "output/pdfs"
 }  # NOTE: Must have resolve otherwise you'll have problems
 
-
+# MOCK FUNCS
 @mock.patch.dict(dirs, mock_dirs, clear=True)
 def initialize_test_driver():
-    """ By mocking the directory paths we force webdriver to set test/output/pdfs as default download directory"""
+    """ By mocking the directory paths we force webdriver to set
+    test/output/pdfs as default download directory """
     return initialize_driver()
-
 
 # TESTS
 class TestPdfDownload(unittest.TestCase):
     def setUp(self) -> None:
+        # create clean output folder
+        if mock_dirs["pdfs"].is_dir():
+            rmtree(mock_dirs["pdfs"])
+        mock_dirs["pdfs"].mkdir(parents=True, exist_ok=True)
+        # init driver
         self.driver = initialize_test_driver()
-        # self.driver = initialize_driver()
-        mock_dirs["pdfs"].mkdir(parents=True, exist_ok=True)  # make directory
-        logs_config(paths["logs_config_test"])
+
 
     def tearDown(self) -> None:
-        rmtree(mock_dirs["pdfs"])
+        pass
 
     @mock.patch.dict(dirs, mock_dirs, clear=True)
     def test_download_pdf(self):
